avdtp_util: fix store_bit16 issue #233

+++ src/classic/avdtp_util.c
@@ -204,7 +204,7 @@ int get_bit16(uint16_t bitmap, int position){
     return (bitmap >> position) & 1;
 }
 
-uint8_t store_bit16(uint16_t bitmap, int position, uint8_t value){
+uint16_t store_bit16(uint16_t bitmap, int position, uint8_t value){
     if (value){
         bitmap |= 1 << position;
     } else {
